DeviceSettings: Play nice with tianma
Implement an alternative way to generate the default color values and the final preset values by asking if the device has a tianma screen. Which are prone to have burn-in displays.

Let's make that all the values respect this limitation of using 90% of the RGB values

ElDainosor 21 edit: Updated to be more reliable when applying values and not setting everything to 0.