[DO NOT MERGE] DeviceSettings: Switch from KCAL to ColorDisplayManager
Since kdrag0n added support for RGB values, we're going to take our KCAL
changes and make it work on the system ColorDisplayManager.
This makes the system color modifications (like Night Light) work
without setting the screen to the max values. Even our tianma changes work.
There might be a better implementation of this for sure. Or a way
to remove the settings seekbars with an overlay.

Also, this makes presets work, at least in a limited way.